c: binary tree logic
Binary tree thumb rule
=================================
will not have any duplicate data
Inorder traversal will always give the sorted order.
if the data is lesser than than the root node it is inserted
in the right side of the node.
if the data is greater than the root nide it is inserted in
the left side of the node.

Signed-off-by: WebEm <embconnect001@gmail.com>