c: Add c prog to find number of bits set in integer
Code to find the no_of_bits set in an integer using
Brian Kernighan's algorithm.

Brian Kernighan's algorithm is used to find the number
of set bits in a number. The idea behind the algorithm
is that when we subtract one from an integer, all the
bits following the rightmost set of bits are inverted,
turning 1 to 0 and 0 to 1. The rightmost set bit also
gets inverted with the bits right to it.

Signed-off-by: Webem <embconnect001@gmail.com>